notifications/event-onclick-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - onclick (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<meta name="flags" content="interact">
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p>Test step:</p>
<ol>
  <li>Click the "Notify me!" button.</lu>
  <li>Click the notification's body area in 60 seconds.</lu>
</ol>
<input value="Notify me!" type="button" onclick="notifyMe();"/>
<div id="log"></div>
<script>
  t = async_test("the notification.onclick event can be invoked");

  t.step(function () {
    assert_true("Notification" in window, "Notification exists");  
  });

  var notification;
  function notifyMe() {
    if(Notification.permission === "granted") {
      notification = new Notification("New notification", {
        body: "click me"
      });
      clickNotification();
    } else if(Notification.permission !== 'denied') {
      Notification.requestPermission(function (permission) {
        if (!('permission' in Notification)) {
          Notification.permission = permission;
        }

        if (permission === "granted") {
          notification = new Notification("New notification", {
            body: "click me"
          });
          clickNotification();
        }
      });
    }
  }

  function clickNotification() {
    notification.onclick = t.step_func(function(e) {
      t.done();
    });
  }

</script>
